Create GUI prototype

+++ src/simu/framework/IMoottori.java
@@ -4,7 +4,7 @@ public interface IMoottori {
 
 	// Kontrolleri käyttää tätä rajapintaa
 
-	public void setSimulointiaika(double aika);
+	public void setSimulointiaika(long aika);
 
 	public void setViive(long aika);
