+++ src/simu/framework/Saapumisprosessi.java
@@ -16,7 +16,7 @@ public class Saapumisprosessi {
 	}
 
 	public void generoiSeuraava() {
-		Tapahtuma t = new Tapahtuma(tyyppi, Kello.getInstance().getAika() + generaattori.sample());
+		Tapahtuma t = new Tapahtuma(tyyppi, Kello.getInstance().getAika() + (long) generaattori.sample());
 		tapahtumalista.lisaa(t);
 	}
